git-svn-id: http://svn.dietrich-hosting.de/svn/bashor/trunk@68 ed1cd6a4-a050-4899-83da-e42ea41b31cf

+++ classes/Terminal/Graphic.sh
@@ -28,30 +28,24 @@
 # $?    0:OK    1:ERROR
 function CLASS_Terminal_Graphic_setPixel()
 {
-    optSetOpts "BUX";
-    optSetArgs "$@";
-    
+    : ${1:?};
+    : ${2:?};
     loadClassOnce 'Terminal';
 
-    argIsNotEmpty 1 || error "Param 1 not set";
-    argIsNotEmpty 2 || error "Param 2 not set";
-
     class Terminal saveCurser;
-    class Terminal moveCurserBy "`argValue 1`" "`argValue 2`";
-    argIsNotEmpty 4 && class Terminal setBackgroundColorAnsi "`argValue 4`";
-    argIsNotEmpty 5 && class Terminal setFordergroundColorAnsi "`argValue 5`";
-    
-    if [ "$1" != '--' ]; then
-        optIsset 'B' && class Terminal setStyleBold 1;
-        optIsset 'U' && class Terminal setStyleUnderline 1;
-        optIsset 'X' && class Terminal setExtendedCharacters 1;
+    class Terminal moveCurserBy "$1" "$2";
+    [ -n "$4" ] && class Terminal setBackgroundColorAnsi "$4";
+    [ -n "$5" ] && class Terminal setFordergroundColorAnsi "$5";
+    if [ -n "$6" ]; then
+        echo "$6" | grep 'B' -q && class Terminal setStyleBold 1;
+        echo "$6" | grep 'U' -q && class Terminal setStyleUnderline 1;
+        echo "$6" | grep 'X' -q && class Terminal setExtendedCharacters 1;
     fi
     
-    local char="`argValue 3`";
-    local char="${char:- }";
+    local char="${3:- }";
     echo -n "${char:0:1}";
-    if [ "$1" != '--' ]; then
-        optIsset 'X' && class Terminal setExtendedCharacters 0;
+    if [ -n "$6" ]; then
+        echo "$6" | grep 'X' -q && class Terminal setExtendedCharacters 0;
     fi
     class Terminal resetStyle;
     class Terminal restoreCurser;
@@ -71,29 +65,24 @@ function CLASS_Terminal_Graphic_setPixel()
 # $?    0:OK    1:ERROR
 function CLASS_Terminal_Graphic_printText()
 {
-    optSetOpts "BUX";
-    optSetArgs "$@";
-
-    argIsNotEmpty 1 || error "Param 1 not set";
-    argIsNotEmpty 2 || error "Param 2 not set";
-    argIsNotEmpty 3 || error "Param 3 not set";
-
+    : ${1:?};
+    : ${2:?};
+    : ${3:?};
     loadClassOnce 'Terminal';
 
     class Terminal saveCurser;
-    class Terminal moveCurserBy "`argValue 1`" "`argValue 2`";
-    argIsNotEmpty 4 && class Terminal setBackgroundColorAnsi "`argValue 4`";
-    argIsNotEmpty 5 && class Terminal setFordergroundColorAnsi "`argValue 5`";    
-    
-    if [ "$1" != '--' ]; then
-        optIsset 'B' && class Terminal setStyleBold 1;
-        optIsset 'U' && class Terminal setStyleUnderline 1;
-        optIsset 'X' && class Terminal setExtendedCharacters 1;
+    class Terminal moveCurserBy "$1" "$2";
+    [ -n "$4" ] && class Terminal setBackgroundColorAnsi "$4";
+    [ -n "$5" ] && class Terminal setFordergroundColorAnsi "$5";
+    if [ -n "$6" ]; then
+        echo "$6" | grep 'B' -q && class Terminal setStyleBold 1;
+        echo "$6" | grep 'U' -q && class Terminal setStyleUnderline 1;
+        echo "$6" | grep 'X' -q && class Terminal setExtendedCharacters 1;
     fi
     
-    echo -n "`argValue 3`";
-    if [ "$1" != '--' ]; then
-        optIsset 'X' && class Terminal setExtendedCharacters 0;
+    echo -n "$3";
+    if [ -n "$6" ]; then
+        echo "$6" | grep 'X' -q && class Terminal setExtendedCharacters 0;
     fi
     class Terminal resetStyle;
     class Terminal restoreCurser;
@@ -114,37 +103,32 @@ function CLASS_Terminal_Graphic_printText()
 # $?    0:OK    1:ERROR
 function CLASS_Terminal_Graphic_printRectangleFilled()
 {
-    optSetOpts "BUX";
-    optSetArgs "$@";
-
-    argIsNotEmpty 1 || error "Param 1 not set";
-    argIsNotEmpty 2 || error "Param 2 not set";
-    argIsNotEmpty 3 || error "Param 3 not set";
-    argIsNotEmpty 4 || error "Param 4 not set";
-    argIsset 5 || error "Param 5 not set";
-
+    : ${1:?};
+    : ${2:?};
+    : ${3:?};
+    : ${4:?};
     loadClassOnce 'Terminal';
     loadClassOnce 'Escape';
 
     (
-        local char="`argValue 5`";
+        local char="$5";
         local char="${char:- }";
         local char="${char:0:1}";
         local char=`class Escape regExReplacement "$char" '/'`;
         
-        local h="`argValue 4`";
+        local h="$4";
         class Terminal saveCurser;
         
-        argIsNotEmpty 6 && class Terminal setBackgroundColorAnsi "`argValue 6`";
-        argIsNotEmpty 7 && class Terminal setFordergroundColorAnsi "`argValue 7`";
-        if [ "$1" != '--' ]; then
-            optIsset 'B' && class Terminal setStyleBold 1;
-            optIsset 'U' && class Terminal setStyleUnderline 1;
-            optIsset 'X' && class Terminal setExtendedCharacters 1;
+        [ -n "$6" ] && class Terminal setBackgroundColorAnsi "$6";
+        [ -n "$7" ] && class Terminal setFordergroundColorAnsi "$7";
+        if [ -n "$8" ]; then
+            echo "$8" | grep 'B' -q && class Terminal setStyleBold 1;
+            echo "$8" | grep 'U' -q && class Terminal setStyleUnderline 1;
+            echo "$8" | grep 'X' -q && class Terminal setExtendedCharacters 1;
         fi
         
-        class Terminal moveCurserBy "`argValue 1`" "`argValue 2`";
-        local v3="`argValue 3`";
+        class Terminal moveCurserBy "$1" "$2";
+        local v3="$3";
         local tmp=`
             dd if=/dev/zero 2>/dev/null bs="$v3" count="1" | tr '\0' "0" | sed 's/0/'"$char"'/g';
             class Terminal moveCurserBy -"$v3" 1;
@@ -152,9 +136,10 @@ function CLASS_Terminal_Graphic_printRectangleFilled()
         local out='';
         for i in `seq "$h"`; do local out="${out}${tmp}"; done;
         echo -n "$out";
-        if [ "$1" != '--' ]; then
-            optIsset 'X' && class Terminal setExtendedCharacters 0;
+        if [ -n "$8" ]; then
+            echo "$8" | grep 'X' -q && class Terminal setExtendedCharacters 0;
         fi
+        
         class Terminal resetStyle;    
         class Terminal restoreCurser;
     )
